test(eve): isolate task reporting from sandboxes
