--- .changeset/stable-sandbox-fanout.md
@@ -1,5 +1,0 @@
----
-"eve": patch
----
-
-Reattach to a named Vercel Sandbox when concurrent creates race so parallel agents can share an inherited sandbox reliably.
