+++ e2e/fixtures/agent-task-reporting/agent/sandbox.ts
@@ -1,0 +1,10 @@
+import { defineSandbox } from "eve/sandbox";
+import { justbash } from "eve/sandbox/just-bash";
+
+// This eval exercises background-task delivery, not shell or filesystem behavior.
+// Keep external sandbox provisioning out of the experiment.
+export default defineSandbox({
+  backend: justbash({
+    filesystem: ({ justBash }) => new justBash.InMemoryFs(),
+  }),
+});
