+++ e2e/fixtures/agent-task-reporting/package.json
@@ -14,6 +14,7 @@
     "@eve-e2e/config": "workspace:*",
     "@workflow/world-postgres": "catalog:",
     "eve": "workspace:*",
+    "just-bash": "3.1.0",
     "zod": "catalog:"
   },
   "devDependencies": {
